Trasform score from double to integer and bug side effects

+++ app/src/main/java/wazzle/model/common/BonusManager.java
@@ -52,9 +52,9 @@ public interface BonusManager {
 	 * 
 	 * @param currentScore the current score obtained by the user.
 	 * @param gridTotalScore the sum of all letter scores in the grid.
-	 * @return double which represents the updated score.
+	 * @return int which represents the updated score.
 	 */
-	double applyScoreBonus(final double currentScore, final double gridTotalScore);
+	int applyScoreBonus(final int currentScore, final int gridTotalScore);
 	
 	/**
 	 * Apply the implications of the WordBonus.
